+++ crates/ruma-common/src/identifiers/matrix_uri.rs
@@ -32,6 +32,9 @@ pub enum MatrixId {
     User(OwnedUserId),
 
     /// An event ID.
+    ///
+    /// Constructing this variant from an `OwnedRoomAliasId` is deprecated, because room aliases
+    /// are mutable, so the URI might break after a while.
     Event(OwnedRoomOrAliasId, OwnedEventId),
 }
 
@@ -572,12 +575,11 @@ mod tests {
                 .to_string(),
             "https://matrix.to/#/!ruma:notareal.hs?via=notareal.hs"
         );
-        assert_eq!(
-            room_alias_id!("#ruma:notareal.hs")
-                .matrix_to_event_uri(event_id!("$event:notareal.hs"))
-                .to_string(),
-            "https://matrix.to/#/%23ruma:notareal.hs/$event:notareal.hs"
-        );
+        #[allow(deprecated)]
+        let uri = room_alias_id!("#ruma:notareal.hs")
+            .matrix_to_event_uri(event_id!("$event:notareal.hs"))
+            .to_string();
+        assert_eq!(uri, "https://matrix.to/#/%23ruma:notareal.hs/$event:notareal.hs");
         assert_eq!(
             room_id!("!ruma:notareal.hs")
                 .matrix_to_event_uri(event_id!("$event:notareal.hs"))
@@ -869,12 +871,11 @@ mod tests {
                 .to_string(),
             "matrix:roomid/ruma:notareal.hs?via=notareal.hs&via=anotherunreal.hs&action=join"
         );
-        assert_eq!(
-            room_alias_id!("#ruma:notareal.hs")
-                .matrix_event_uri(event_id!("$event:notareal.hs"))
-                .to_string(),
-            "matrix:r/ruma:notareal.hs/e/event:notareal.hs"
-        );
+        #[allow(deprecated)]
+        let uri = room_alias_id!("#ruma:notareal.hs")
+            .matrix_event_uri(event_id!("$event:notareal.hs"))
+            .to_string();
+        assert_eq!(uri, "matrix:r/ruma:notareal.hs/e/event:notareal.hs");
         assert_eq!(
             room_id!("!ruma:notareal.hs")
                 .matrix_event_uri(event_id!("$event:notareal.hs"))