+++ crates/ruma-common/CHANGELOG.md
@@ -23,6 +23,8 @@ Improvements:
 - Add `MatrixVersion::V1_11`
 - Clarify in the docs of `AuthScheme` that sending an access token via a query
   parameter is deprecated, according to MSC4126 / Matrix 1.11.
+- Constructing a Matrix URI for an event with a room alias is deprecated,
+  according to MSC4132 / Matrix 1.11
 
 # 0.13.0